Moving trace test too
