+++ opentelemetry-api/tests/trace/test_globals.py
@@ -1,9 +1,7 @@
 import unittest
-from logging import WARNING
 from unittest.mock import patch
 
 from opentelemetry import context, trace
-from opentelemetry.sdk.trace import TracerProvider  # type:ignore
 
 
 class TestGlobals(unittest.TestCase):
@@ -21,23 +19,6 @@ class TestGlobals(unittest.TestCase):
         mock_provider = unittest.mock.Mock()
         trace.get_tracer("foo", "var", mock_provider)
         mock_provider.get_tracer.assert_called_with("foo", "var")
-
-    def test_tracer_provider_override_warning(self):
-        """trace.set_tracer_provider should throw a warning when overridden"""
-        trace.set_tracer_provider(TracerProvider())
-        tracer_provider = trace.get_tracer_provider()
-        with self.assertLogs(level=WARNING) as test:
-            trace.set_tracer_provider(TracerProvider())
-            self.assertEqual(
-                test.output,
-                [
-                    (
-                        "WARNING:opentelemetry.trace:Overriding of current "
-                        "TracerProvider is not allowed"
-                    )
-                ],
-            )
-        self.assertIs(tracer_provider, trace.get_tracer_provider())
 
 
 class TestTracer(unittest.TestCase):
